requirements.txt
docopt==0.6.2
Jinja2==2.10
Pillow==5.0.0
schema==0.6.7
selenium==3.9.0
